requirements.txt
fastapi
uvicorn
requests
python-telegram-bot==20.6
psycopg2-binary
asyncpg
pandas
numpy
